code/ch02/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.0
cycler==0.12.1
easyocr==1.7.2
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.4
numpy==2.0.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pyparsing==3.2.3
pyproj==3.6.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
six==1.17.0
sympy==1.14.0
tifffile==2024.8.30
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.138
ultralytics-thop==2.0.14
urllib3==2.4.0
zipp==3.21.0
